выполнено 4.1 аватарки не оторажаются
