+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -1,6 +1,7 @@
 package ru.hogwarts.school.controller;
 
 
+import java.awt.*;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -45,6 +46,21 @@ public class StudentController {
         }
         return ResponseEntity.ok(student);
     }
+
+    @GetMapping("/amount-all-student")
+    public Long getAmountOfAllStudents(){
+        return studentService.getAmountOfAllStudents();
+    }
+
+    @GetMapping("/average-age-of-all-students")
+    public double getAverageAgeOfAllStudents(){
+        return studentService.getAverageAgeOfAllStudents();
+    }
+    @GetMapping("/get-last-five-students()")
+    public Collection<Student> getLastFiveStudents(){
+        return studentService.getLastFiveStudents();
+    }
+
     @PostMapping
     public Student createStudent (@RequestBody Student student) {
         return studentService.addStudent(student);
@@ -81,39 +97,5 @@ public class StudentController {
     public Faculty getFacultyByStudent(@PathVariable long id){
         return studentService.getFacultyByStudent(id);}
 
-    @PostMapping(value = "/{id}/avatar", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<String> uploadAvatar(@PathVariable Long id, @RequestParam MultipartFile avatar) throws IOException {
-        if (avatar.getSize() > 1024 * 300) {
-            return ResponseEntity.badRequest().body("File is too big");
-        }
 
-        studentService.uploadAvatar(id, avatar);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/{id}/avatar/preview")
-    public ResponseEntity<byte[]> downloadAvatar(@PathVariable Long id) {
-        Avatar avatar = studentService.findAvatar(id);
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.parseMediaType(avatar.getMediaType()));
-        headers.setContentLength(avatar.getData().length);
-
-        return ResponseEntity.status(HttpStatus.OK).headers(headers).body(avatar.getData());
-    }
-
-    @GetMapping(value = "/{id}/avatar")
-    public void downloadAvatar(@PathVariable Long id, HttpServletResponse response) throws IOException {
-        Avatar avatar = studentService.findAvatar(id);
-
-        Path path = Path.of(avatar.getFilePath());
-
-        try (InputStream is = Files.newInputStream(path);
-             OutputStream os = response.getOutputStream();) {
-            response.setStatus(200);
-            response.setContentType(avatar.getMediaType());
-            response.setContentLength((int) avatar.getFileSize());
-            is.transferTo(os);
-        }
-    }
 }
