+++ src/main/java/ru/hogwarts/school/service/AvatarService.java
@@ -1,0 +1,80 @@
+package ru.hogwarts.school.service;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import ru.hogwarts.school.model.Avatar;
+import ru.hogwarts.school.model.Student;
+import ru.hogwarts.school.repository.AvatarRepository;
+import ru.hogwarts.school.repository.StudentRepository;
+
+import javax.transaction.Transactional;
+import java.io.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Collection;
+
+import static java.nio.file.StandardOpenOption.CREATE_NEW;
+
+@Service
+@Transactional
+public class AvatarService {
+
+    @Value("${path.to.avatars.folder}")
+    private String avatarsDir;
+
+    private StudentRepository studentRepository;
+    private final AvatarRepository avatarRepository;
+
+    private final StudentService studentService;
+
+    public AvatarService(StudentRepository studentRepository, AvatarRepository avatarRepository, StudentService studentService) {
+
+        this.studentRepository = studentRepository;
+        this.avatarRepository = avatarRepository;
+        this.studentService = studentService;
+    }
+
+    public Avatar findAvatar(long studentId) {
+        return avatarRepository.findByStudentId(studentId).orElseThrow();
+    }
+
+    public void uploadAvatar(Long studentId, MultipartFile file) throws IOException {
+        Student student = studentService.findStudent(studentId);
+
+        Path filePath = Path.of(avatarsDir, studentId + "." + getExtension(file.getOriginalFilename()));
+        Files.createDirectories(filePath.getParent());
+        Files.deleteIfExists(filePath);
+
+        try (InputStream is = file.getInputStream();
+             OutputStream os = Files.newOutputStream(filePath, CREATE_NEW);
+             BufferedInputStream bis = new BufferedInputStream(is, 1024);
+             BufferedOutputStream bos = new BufferedOutputStream(os, 1024);
+        ) {
+            bis.transferTo(bos);
+        }
+
+        Avatar avatar = avatarRepository.findByStudentId(studentId).orElseGet(Avatar::new);
+        avatar.setStudent(student);
+        avatar.setFilePath(filePath.toString());
+        avatar.setFileSize(file.getSize());
+        avatar.setMediaType(file.getContentType());
+        avatar.setData(file.getBytes());
+
+        avatarRepository.save(avatar);
+    }
+    private String getExtension(String fileName) {
+        return fileName.substring(fileName.lastIndexOf(".") + 1);
+    }
+
+    public ResponseEntity<Collection<Avatar>> getAll(Integer pageNumber, Integer pageSize) {
+        PageRequest pageRequest = PageRequest.of(pageNumber-1, pageSize);
+        Collection<Avatar> avtarList = avatarRepository.findAll(pageRequest).getContent();
+        if (avtarList.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(avtarList);
+    }
+}
