+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -1,53 +1,22 @@
 package ru.hogwarts.school.service;
 
-import static java.nio.file.StandardOpenOption.CREATE_NEW;
 
-
-import java.awt.Graphics2D;
-import java.awt.image.BufferedImage;
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Collection;
-import javax.imageio.ImageIO;
-import javax.persistence.EntityNotFoundException;
-import javax.transaction.Transactional;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-import ru.hogwarts.school.model.Avatar;
 import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.model.Student;
-import ru.hogwarts.school.repository.AvatarRepository;
 import ru.hogwarts.school.repository.StudentRepository;
 
-import java.io.IOException;
-import java.util.Collection;
-
 @Service
-@Transactional
 public class StudentService {
 
-    @Value("${path.to.avatars.folder}")
-    private String avatarsDir;
-
     private StudentRepository studentRepository;
-    private final AvatarRepository avatarRepository;
 
-    public StudentService(StudentRepository studentRepository, AvatarRepository avatarRepository) {
-
+    public StudentService(StudentRepository studentRepository) {
         this.studentRepository = studentRepository;
-        this.avatarRepository = avatarRepository;
     }
+
 
     public Student addStudent(Student student) {
         return studentRepository.save(student);
@@ -73,36 +42,15 @@ public class StudentService {
         return findStudent(id).getFaculty();
         
     }
-    public Avatar findAvatar(long studentId) {
-        return avatarRepository.findByStudentId(studentId).orElseThrow();
+    public Long getAmountOfAllStudents(){
+        return studentRepository.getAmountOfAllStudents();
+
+    }
+    public Double getAverageAgeOfAllStudents(){
+        return studentRepository.getAverageAgeOfAllStudents();
     }
 
-    public void uploadAvatar(Long studentId, MultipartFile file) throws IOException {
-        Student student = findStudent(studentId);
-
-        Path filePath = Path.of(avatarsDir, studentId + "." + getExtension(file.getOriginalFilename()));
-        Files.createDirectories(filePath.getParent());
-        Files.deleteIfExists(filePath);
-
-        try (InputStream is = file.getInputStream();
-             OutputStream os = Files.newOutputStream(filePath, CREATE_NEW);
-             BufferedInputStream bis = new BufferedInputStream(is, 1024);
-             BufferedOutputStream bos = new BufferedOutputStream(os, 1024);
-        ) {
-            bis.transferTo(bos);
-        }
-
-        Avatar avatar = avatarRepository.findByStudentId(studentId).orElseGet(Avatar::new);
-        avatar.setStudent(student);
-        avatar.setFilePath(filePath.toString());
-        avatar.setFileSize(file.getSize());
-        avatar.setMediaType(file.getContentType());
-        avatar.setData(file.getBytes());
-
-        avatarRepository.save(avatar);
-    }
-
-    private String getExtension(String fileName) {
-        return fileName.substring(fileName.lastIndexOf(".") + 1);
+    public Collection<Student> getLastFiveStudents(){
+        return  studentRepository.getLastFiveStudents();
     }
 }
